Begin breaking component out to work with architecture
